style: update style for header
