+++ src/components/header/bottom-header/bottom-header.module.css
@@ -4,11 +4,12 @@
   align-items: center;
   width: 100%;
   background-color: var(--clr-bg);
+  padding: 0 63px;
 }
 
 .bottomNavigationMenu {
   display: flex;
-  padding: 14px 14px;
+  padding: 14px 0;
 }
 
 .input {
