+++ src/components/header/bottom-header/left-menu/left-menu.tsx
@@ -1,21 +1,11 @@
 import styles from '../../header.module.css';
-import svgStyles from '../../../svg-builder/svg.module.css';
 import { NavLink } from 'react-router-dom';
 import { leftLinks } from '../constants';
-import { PagePath } from '../../../../router/enums';
-import { messages } from '../messages';
-import { SvgBuilder } from '../../../svg-builder/svg-builder';
-import { IconType } from '../../../svg-builder/enums';
 import classNames from 'classnames';
 
 export const LeftNavMenu = () => {
   return (
     <nav className={styles.navigationMenu}>
-      <NavLink to={PagePath.root}>
-        {messages.defaultLink}
-        <SvgBuilder iconType={IconType.Logo} className={svgStyles.large} />
-      </NavLink>
-
       {leftLinks.map(link => {
         return (
           <NavLink
