--- src/components/header/top-header/messages.ts
@@ -1,8 +1,0 @@
-export const messages = {
-  loginLink: 'Login',
-  logoutLink: 'Logout',
-  registerLink: 'Registration',
-  emailLink: 'khnizovad@gmail.com',
-  phoneLink: '8 707 275 4341',
-  basket: '',
-};
