+++ src/components/header/top-header/right-menu/messages.ts
@@ -1,0 +1,4 @@
+export const messages = {
+  loginLink: 'Login',
+  registerLink: 'Sign up',
+};
