+++ src/components/header/top-header/right-menu/right-nav-menu.tsx
@@ -1,7 +1,7 @@
 import styles from '../../header.module.css';
 import svgStyles from '../../../svg-builder/svg.module.css';
 import { NavLink } from 'react-router-dom';
-import { rightLinks } from '../constants';
+import { rightLinks } from './constants';
 import { SvgBuilder } from '../../../svg-builder/svg-builder';
 
 export const RightNavMenu = () => {
@@ -10,8 +10,8 @@ export const RightNavMenu = () => {
       {rightLinks.map(link => {
         return (
           <NavLink key={link.label} to={link.to} className={styles.link}>
-            {link.label}
             <SvgBuilder iconType={link.iconType} className={svgStyles.small} />
+            {link.label}
           </NavLink>
         );
       })}
