+++ src/components/header/top-header/top-header.module.css
@@ -1,7 +1,10 @@
 .topHeader {
+  position: relative;
   display: flex;
   justify-content: space-between;
   align-items: center;
   width: 100%;
+  height: 50px;
   background-color: var(--clr-accent-2);
+  padding: 0 63px;
 }
