+++ src/components/svg-builder/svg.module.css
@@ -14,12 +14,8 @@
   height: 20px;
 }
 
-.large {
-  width: 60px;
-  height: 60px;
+.logo {
+  width: 40px;
+  height: 30px;
 }
 
-.basket {
-  width: 24px;
-  height: 24px;
-}
