+++ src/components/wrapper/wrapper.module.css
@@ -1,4 +1,7 @@
 .wrapper {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
   max-width: 1440px;
   margin: 0 auto;
 }
