+++ src/global.module.css
@@ -45,7 +45,9 @@ a:active {
   --clr-disabled-300: #a0a0a0;
   --clr-disabled-600: #5f5f5f;
 
-  --clr-bg: #ffffff;
+  --clr-bg-1: #ffffff;
+  --clr-bg-2: #F6F5FF;
+
   --clr-line: #000000;
 }
